app/requirements.txt
aldryn-apphooks-config==0.6.0
asgiref==3.6.0
cssselect2==0.7.0
Django==4.2.6
django-admin-sortable==2.3
django-appdata==0.3.2
django-classy-tags==4.0.0
django-cms==3.11.4
django-colorfield==0.7.2
django-crispy-forms==1.14.0
django-filer==3.1.0
django-filter==22.1
django-formtools==2.4
django-js-asset==2.0.0
django-mptt==0.13.4
django-ordered-model==3.6
django-polymorphic==3.1.0
django-recaptcha==3.0.0
django-sekizai==4.0.0
django-taggit==3.0.0
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.6.1
djangocms-admin-style==3.2.3
djangocms-apphook-setup==0.4.1
djangocms-attributes-field==2.1.0
djangocms-column==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-link==3.1.0
djangocms-picture==4.0.0
djangocms-snippet==3.1.0
djangocms-style==3.0.0
djangocms-text-ckeditor==5.1.1
djangocms-transfer==1.0.1
djangocms-video==3.0.0
easy-thumbnails==2.8.3
html5lib==1.1
lxml==4.9.1
packaging==23.2
Pillow==9.2.0
pkg_resources==0.0.0
pytz==2022.2.1
reportlab==3.6.11
six==1.16.0
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.1.1
Unidecode==1.1.2
webencodings==0.5.1
